# dashboard/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Flaglane API",
    "description" : "Feature flags and remote configuration. `/api/**` serves the dashboard with a JWT; `/sdk/**` serves SDKs with an API key.",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "0.1.0"
  },
...
  "paths" : {
...
    "/api/projects/{projectKey}/flags/{flagKey}/config/{envKey}/overrides" : {
      "get" : {
        "tags" : [ "Targeting" ],
        "summary" : "Get a flag's user overrides in an environment",
        "operationId" : "overrides",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The overrides",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OverridesResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Targeting" ],
        "summary" : "Replace a flag's user overrides in an environment",
        "description" : "The whole set at once. Overrides apply to server keys only: a client key's ruleset carries none, because their user keys are real user identifiers (FR-KEY-008).",
        "operationId" : "replaceOverrides",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReplaceOverridesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "409" : {
            "description" : "The flag is archived",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "The overrides as stored",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OverridesResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/api/projects/{projectKey}/environments" : {
      "get" : {
        "tags" : [ "Environments" ],
        "summary" : "List a project's environments",
        "operationId" : "list_2",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The project's environments",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EnvironmentResponse"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or environment among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Environments" ],
        "summary" : "Create an environment",
        "description" : "Every flag in the project gets a configuration here, disabled, falling through to `false` at 0% (FR-ENV-004).",
        "operationId" : "create_2",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateEnvironmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnvironmentResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The project already has an environment with this key",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or environment among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/api/projects/{projectKey}/environments/{envKey}" : {
      "delete" : {
        "tags" : [ "Environments" ],
        "summary" : "Delete an environment",
        "description" : "Deletes the environment with its keys, configurations, rules and overrides. Its audit entries are kept. Refused while it holds a key that is not revoked (FR-ENV-003).",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "409" : {
            "description" : "The environment holds an API key that is not revoked",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project or environment among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri"
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : [ "string", "null" ]
          },
          "instance" : {
            "type" : [ "string", "null" ],
            "format" : "uri"
          },
          "properties" : {
            "type" : [ "object", "null" ],
            "additionalProperties" : { }
          }
        }
      },
...
      "OverrideRequest" : {
        "type" : "object",
        "properties" : {
          "userKey" : {
            "type" : "string",
            "example" : "u-1042",
            "maxLength" : 256,
            "minLength" : 0
          },
          "value" : {
            "type" : "boolean"
          }
        },
        "required" : [ "userKey", "value" ]
      },
      "ReplaceOverridesRequest" : {
        "type" : "object",
        "properties" : {
          "overrides" : {
            "type" : "array",
            "description" : "An empty list removes every override (ADR-023: at most 1,000). Overrides do not apply to client keys, on either serving path (FR-KEY-008)",
            "items" : {
              "$ref" : "#/components/schemas/OverrideRequest"
            },
            "maxItems" : 1000,
            "minItems" : 0
          }
        },
        "required" : [ "overrides" ]
      },
      "OverridesResponse" : {
        "type" : "object",
        "properties" : {
          "overrides" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OverrideRequest"
            }
          }
        },
        "required" : [ "overrides" ]
      },
...
      "CreateEnvironmentRequest" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string",
            "description" : "Unique within the project",
            "example" : "qa",
            "pattern" : "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
          },
          "name" : {
            "type" : "string",
            "example" : "QA",
            "maxLength" : 100,
            "minLength" : 0
          }
        },
        "required" : [ "key", "name" ]
      },
      "EnvironmentResponse" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string",
            "example" : "production"
          },
          "name" : {
            "type" : "string",
            "example" : "Production"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "key", "name" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "description" : "An access token from `POST /api/auth/login`, valid for 8 hours.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "sdk-key" : {
        "type" : "http",
        "description" : "An API key issued for one environment. A server key reads every flag; a client key reads client-side-visible flags only, with no user overrides.",
        "scheme" : "bearer",
        "bearerFormat" : "flg_srv_… or flg_cli_…"
      }
    }
  }
}